Resolve the configuration path

Resolve the configuration path

resolve_config_path()

Specifying Config path

The pipetree config path can be set using a number of mechanisms.

If the path is set in multiple places, the path chosen will prioritize in order of more specific settings befroe less specific.

In order of first-used, the config path can be described as below:

Explicitly providing the path in a call to a pipetree function

getOption('pipetree.config')

The system environment variable PIPETREE_CONFIG_PATH

The default config path given by default_config_path().